refactor(tui): split Ghostty terminal runner
