+++ test/README.md
@@ -3,8 +3,8 @@
 This template has two layers of tests:
 
 - `main.zig` keeps fast Zig smoke tests close to the build graph.
-- `test_*_scenarios.py` files are end-to-end terminal scenarios for project users.
-- `test_terminal_harness.py` covers the reusable Python harness itself.
+- The C Ghostty VT runner covers CLI contracts plus PTY-backed TUI flows when libghostty-vt is available.
+- `test_*_scenarios.py` and `test_terminal_harness.py` are the Python fallback for platforms without Ghostty VT.
 
 Run the default Zig suite:
 
@@ -24,18 +24,17 @@ Run the same scenarios against a TUI-enabled build:
 zig build -Denable-tui=true terminal-test
 ```
 
-`zig build terminal-test` always runs the Python-discovered CLI scenario and
-harness unit tests. The default PTY backend is `auto`: it uses the C Ghostty VT
-runner when `libghostty-vt` is available through `pkg-config`, otherwise it
-falls back to the Python harness.
+The default terminal-test backend is `auto`: it uses the C Ghostty VT runner
+when `libghostty-vt` is available through `pkg-config`, otherwise it falls back
+to the Python harness. The Ghostty path does not invoke Python.
 
-The Ghostty VT runner lives in `terminal_vt_runner.c`. It runs the app in a
-pseudo-terminal, feeds output through libghostty-vt, snapshots the screen with
-Ghostty's formatter API, and drives deterministic input plus resize actions.
+The Ghostty VT runner is split across `terminal_vt_*.c` files. It runs CLI
+checks directly, runs TUI checks in a pseudo-terminal, feeds output through
+libghostty-vt, snapshots the screen with Ghostty's formatter API, and drives
+deterministic input plus resize actions.
 
-The Python fallback uses `pexpect` plus `pyte` for TUI scenarios. When Ghostty is
-selected, Python TUI scenarios skip and the C runner provides PTY coverage. The
-Nix dev shell provides the Python fallback dependencies. For the Ghostty backend,
+The Python fallback uses `pexpect` plus `pyte` for TUI scenarios. The Nix dev
+shell provides the Python fallback dependencies. For the Ghostty backend,
 install a libghostty-vt build with the development Terminal and Formatter APIs
 and make it visible through `pkg-config` or
 `-Dghostty-vt-prefix=/path/to/ghostty`.
