+++ test/terminal_vt_session.c
@@ -1,0 +1,286 @@
+#include "terminal_vt.h"
+
+static void ghostty_write_pty(GhosttyTerminal terminal, void *userdata,
+                              const uint8_t *data, size_t len) {
+  (void)terminal;
+  vt_session_t *session = userdata;
+  if (!session || session->pty_write_failed || session->master_fd < 0) {
+    return;
+  }
+
+  if (!write_nonblocking_all(session->master_fd, data, len, PTY_TIMEOUT_MS)) {
+    session->pty_write_failed = true;
+    close_if_open(&session->master_fd);
+  }
+}
+
+bool vt_session_start(vt_session_t *session, const char *binary,
+                      const char *const *args, size_t argc, uint16_t cols,
+                      uint16_t rows) {
+  memset(session, 0, sizeof(*session));
+  session->master_fd = -1;
+  session->cols = cols;
+  session->rows = rows;
+
+  GhosttyTerminalOptions opts = {
+      .cols = cols,
+      .rows = rows,
+      .max_scrollback = 1000,
+  };
+  if (ghostty_terminal_new(NULL, &session->terminal, opts) != GHOSTTY_SUCCESS) {
+    fputs("failed to create Ghostty VT terminal\n", stderr);
+    return false;
+  }
+
+  ghostty_terminal_set(session->terminal, GHOSTTY_TERMINAL_OPT_USERDATA,
+                       session);
+  ghostty_terminal_set(session->terminal, GHOSTTY_TERMINAL_OPT_WRITE_PTY,
+                       (const void *)ghostty_write_pty);
+
+  struct winsize ws = {
+      .ws_row = rows,
+      .ws_col = cols,
+      .ws_xpixel = (unsigned short)(cols * 8),
+      .ws_ypixel = (unsigned short)(rows * 16),
+  };
+
+  char **argv = make_argv(binary, args, argc);
+  if (!argv) {
+    ghostty_terminal_free(session->terminal);
+    session->terminal = NULL;
+    return false;
+  }
+
+  const pid_t pid = forkpty(&session->master_fd, NULL, NULL, &ws);
+  if (pid < 0) {
+    perror("forkpty");
+    free(argv);
+    ghostty_terminal_free(session->terminal);
+    session->terminal = NULL;
+    return false;
+  }
+
+  if (pid == 0) {
+    apply_child_env(NULL, 0, true);
+    execv(binary, argv);
+    perror("execv");
+    _exit(127);
+  }
+
+  session->pid = pid;
+  free(argv);
+  set_nonblocking(session->master_fd);
+  return true;
+}
+
+void vt_session_close(vt_session_t *session) {
+  if (!session) {
+    return;
+  }
+
+  if (!session->child_exited && session->pid > 0) {
+    kill(session->pid, SIGTERM);
+    for (int i = 0; i < 20; i++) {
+      int status = 0;
+      const pid_t waited = waitpid(session->pid, &status, WNOHANG);
+      if (waited == session->pid) {
+        session->child_exited = true;
+        break;
+      }
+      usleep(25000);
+    }
+    if (!session->child_exited) {
+      kill(session->pid, SIGKILL);
+      waitpid(session->pid, NULL, 0);
+      session->child_exited = true;
+    }
+  }
+
+  close_if_open(&session->master_fd);
+  if (session->terminal) {
+    ghostty_terminal_free(session->terminal);
+    session->terminal = NULL;
+  }
+  buffer_free(&session->transcript);
+}
+
+static void vt_session_note_exit(vt_session_t *session, int status) {
+  session->child_exited = true;
+  if (WIFEXITED(status)) {
+    session->exit_code = WEXITSTATUS(status);
+  } else if (WIFSIGNALED(status)) {
+    session->exit_code = 128 + WTERMSIG(status);
+  } else {
+    session->exit_code = -1;
+  }
+}
+
+void vt_drain(vt_session_t *session, int idle_ms) {
+  if (!session || session->pty_write_failed || session->master_fd < 0) {
+    return;
+  }
+
+  const int64_t deadline = monotonic_ms() + idle_ms;
+  while (true) {
+    const int64_t now = monotonic_ms();
+    int wait_ms = 0;
+    if (idle_ms > 0) {
+      wait_ms = now >= deadline ? 0 : (int)(deadline - now);
+      if (wait_ms > 50) {
+        wait_ms = 50;
+      }
+    }
+
+    struct pollfd fd = {.fd = session->master_fd, .events = POLLIN};
+    const int polled = poll(&fd, 1, wait_ms);
+    if (polled <= 0) {
+      break;
+    }
+
+    if (!(fd.revents & (POLLIN | POLLHUP | POLLERR))) {
+      break;
+    }
+
+    bool read_any = false;
+    while (true) {
+      uint8_t chunk[READ_CHUNK];
+      const ssize_t n = read(session->master_fd, chunk, sizeof(chunk));
+      if (n > 0) {
+        read_any = true;
+        buffer_append(&session->transcript, chunk, (size_t)n);
+        ghostty_terminal_vt_write(session->terminal, chunk, (size_t)n);
+        continue;
+      }
+      if (n == 0) {
+        close_if_open(&session->master_fd);
+        break;
+      }
+      if (errno == EIO) {
+        close_if_open(&session->master_fd);
+        break;
+      }
+      if (errno == EAGAIN || errno == EWOULDBLOCK || errno == EINTR) {
+        break;
+      }
+      close_if_open(&session->master_fd);
+      break;
+    }
+
+    if (!read_any && idle_ms == 0) {
+      break;
+    }
+  }
+
+  if (!session->child_exited && session->pid > 0) {
+    int status = 0;
+    const pid_t waited = waitpid(session->pid, &status, WNOHANG);
+    if (waited == session->pid) {
+      vt_session_note_exit(session, status);
+    }
+  }
+}
+
+static char *vt_snapshot_text(vt_session_t *session) {
+  GhosttyFormatterTerminalOptions opts =
+      GHOSTTY_INIT_SIZED(GhosttyFormatterTerminalOptions);
+  opts.emit = GHOSTTY_FORMATTER_FORMAT_PLAIN;
+  opts.trim = true;
+
+  GhosttyFormatter formatter = NULL;
+  if (ghostty_formatter_terminal_new(NULL, &formatter, session->terminal,
+                                     opts) != GHOSTTY_SUCCESS) {
+    return NULL;
+  }
+
+  uint8_t *raw = NULL;
+  size_t len = 0;
+  if (ghostty_formatter_format_alloc(formatter, NULL, &raw, &len) !=
+      GHOSTTY_SUCCESS) {
+    ghostty_formatter_free(formatter);
+    return NULL;
+  }
+
+  char *text = calloc(len + 1, 1);
+  if (text) {
+    memcpy(text, raw, len);
+  }
+  ghostty_free(NULL, raw, len);
+  ghostty_formatter_free(formatter);
+  return text;
+}
+
+bool vt_expect_text(vt_session_t *session, const char *needle, int timeout_ms,
+                    char **last_snapshot) {
+  const int64_t deadline = monotonic_ms() + timeout_ms;
+  while (monotonic_ms() <= deadline) {
+    vt_drain(session, 50);
+    if (session->pty_write_failed) {
+      break;
+    }
+    char *snapshot = vt_snapshot_text(session);
+    if (snapshot && contains_text(snapshot, needle)) {
+      if (last_snapshot) {
+        free(*last_snapshot);
+        *last_snapshot = snapshot;
+      } else {
+        free(snapshot);
+      }
+      return true;
+    }
+    if (last_snapshot) {
+      free(*last_snapshot);
+      *last_snapshot = snapshot;
+    } else {
+      free(snapshot);
+    }
+    if (session->child_exited) {
+      break;
+    }
+  }
+  return false;
+}
+
+bool vt_send(vt_session_t *session, const char *bytes) {
+  if (session->pty_write_failed) {
+    return false;
+  }
+  const size_t len = strlen(bytes);
+  if (!write_nonblocking_all(session->master_fd, bytes, len, PTY_TIMEOUT_MS)) {
+    session->pty_write_failed = true;
+    return false;
+  }
+  usleep(50000);
+  vt_drain(session, 0);
+  return true;
+}
+
+bool vt_resize(vt_session_t *session, uint16_t cols, uint16_t rows) {
+  struct winsize ws = {
+      .ws_row = rows,
+      .ws_col = cols,
+      .ws_xpixel = (unsigned short)(cols * 8),
+      .ws_ypixel = (unsigned short)(rows * 16),
+  };
+  if (ioctl(session->master_fd, TIOCSWINSZ, &ws) != 0) {
+    return false;
+  }
+  kill(session->pid, SIGWINCH);
+  session->cols = cols;
+  session->rows = rows;
+  return ghostty_terminal_resize(session->terminal, cols, rows, 8, 16) ==
+         GHOSTTY_SUCCESS;
+}
+
+int vt_wait_for_exit(vt_session_t *session, int timeout_ms) {
+  const int64_t deadline = monotonic_ms() + timeout_ms;
+  while (monotonic_ms() <= deadline) {
+    vt_drain(session, 50);
+    if (session->pty_write_failed) {
+      return -1;
+    }
+    if (session->child_exited) {
+      return session->exit_code;
+    }
+  }
+  return -1;
+}
